🐛 Fix bug getLocalité helper
