+++ .demo/exempleCandidatsPPE2.csv
@@ -190,7 +190,7 @@ Commentaires";Productible annuel (MWh/an);Facteur de charges (kWh/kWc);Montant e
 (AO sol)";"Codes cas
 (AO sol)";"Codes cas
 (AO sol)";Nom projet (doublon);CP (doublon);Commune (doublon);Modifications;Statut;Commentaires;;Type de modification 1;Date de modification 1;Colonne concernée 1;Ancienne valeur 1;Type de modification 2;Date de modification 2;Colonne concernée 2;Ancienne valeur 2;Type de modification 3;Date de modification 3;Colonne concernée 3;Ancienne valeur 3;Type de modification 4;Date de modification 4;Colonne concernée 4;Ancienne valeur 4;Type de modification 5;Date de modification 5;Colonne concernée 5;Ancienne valeur 5
-test-51;PPE2 - Bâtiment;1;;SD3B;Centrale PV 51;1;;1;2.615;1;500;63.91;3.57;;67.48;SD3B;Personne morale;;;PME;DGEC;Ile-de-France;58 rue de Sèvres 92100 Boulogne-Billancourt;Valentin COGNITO;Gérant;Jean-Luc LASSUS;SD3B;Directeur du développement;1 place carpeaux;autre-porteur@test.test;;Centrale PV 51;;;1 place carpeaux;192800;PUTEAUX;Hauts de Seine;Ile-de-France;;2.615;;;;;;;1;500;Non;;;Non;;Déclaration préalable de travaux;;;;;;;;;;;;Bâtiment;1025;;;;;;;;;;;;;;;;oui;Bâtiment;;;;Favorable;Kbis;;Favorable;;;Favorable;Déclaration préalable de travaux;;;AMD;Favorable;GPD;78455;30001.91205;21/10/2021;Favorable;;Favorable;Non;0%;0;;Non;0%;;Favorable;oui;Favorable;;Favorable;;Retenu;Classé;;1;1;1;1;1;;;;N/A;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;;;;;;;;;;;;;;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;2.615;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;Oui;Non;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+test-51;PPE2 - Bâtiment;1;;SD3B;Centrale PV 51;1;;1;2.615;1;500;63.91;3.57;;67.48;SD3B;Personne morale;;;PME;DGEC;Ile-de-France;58 rue de Sèvres 92100 Boulogne-Billancourt;Valentin COGNITO;Gérant;Jean-Luc LASSUS;SD3B;Directeur du développement;1 place carpeaux;autre-porteur@test.test;;Centrale PV 51;;;1 place carpeaux;92800;PUTEAUX;Hauts de Seine;Ile-de-France;;2.615;;;;;;;1;500;Non;;;Non;;Déclaration préalable de travaux;;;;;;;;;;;;Bâtiment;1025;;;;;;;;;;;;;;;;oui;Bâtiment;;;;Favorable;Kbis;;Favorable;;;Favorable;Déclaration préalable de travaux;;;AMD;Favorable;GPD;78455;30001.91205;21/10/2021;Favorable;;Favorable;Non;0%;0;;Non;0%;;Favorable;oui;Favorable;;Favorable;;Retenu;Classé;;1;1;1;1;1;;;;N/A;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;;;;;;;;;;;;;;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;#N/A;2.615;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;Oui;Non;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 test-49;PPE2 - Bâtiment;1;;SD3B;Centrale PV 49;1;;1;1.589;1;424;51.8;9;;62.8;SD3B;Personne morale;;;PME;DGEC;Ile-de-France;58 rue de Sèvres 92100 Boulogne-Billancourt;Valentin COGNITO;Président;Jean-Luc LASSUS;SD3B;Directeur du développement;1 place carpeaux;porteur@test.test;;Centrale PV 49;;;1 rue monge;95019;PARIS;Paris;Ile-de-France;;1.589;;;;;;;1;424;Oui;10%;Au moins 20 personnes physiques;Non;;Permis de construire;;;;;;;; ;;;;Ombrière de parking;1384;;;;;;;;;;;;;;;;oui;Ombrière de parking;;;;Favorable;Kbis;;Favorable;;;Favorable;Permis de construire;;;AMD;Favorable;GPD;47670;30000;21/01/2022;Favorable;;Favorable;Oui;10%;Au moins 20 personnes physiques;;Non;0%;;Favorable;s.o.;Favorable;;Favorable;;Retenu;Classé;;3;1;1;1;1;;;;N/A;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;Silicium monocristallin;123;MODULE PV;0;0;0;0;Chine;0;0;0;0;405;0.211;CELLULE PV;0;0;0;0;Chine;0;0;0;0;PLAQUETTE PV;0;0;0;0;Allemagne;0;0;0;0;POLYSI PV;0;0;0;0;Norvège;0;0;0;0;CONVERT PV;0;0;0;0;Chine;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;;;;;;;;;;;;;;0.021;0.446203;0;0;0;0.05;0.563728;0;0;0;0.0556664;0.104194;1;0;0;0;0;0;0;0;0;0;1;1;1;0;0;0;0;0;1;0;0;0;1;1;"Ces montants sont donnés avec les données disponibles à ce jour pour le porteur de projet. Toutes les informations ne sont pas disponibles.
 
 Le prix des modules comprend l'ensemble des sous-lots (pas de détails disponible pour le candidat)";4.204;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;Non;Non;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
