+++ packages/applications/ssr/src/components/pages/candidature/helpers/index.ts
@@ -1,0 +1,5 @@
+export { getLocalité } from './getLocalité';
+export {
+  getRégionAndDépartementFromCodePostal,
+  type DépartementRégion,
+} from './getRégionAndDépartementFromCodePostal';
